refactor (delivery) : response, request 패키지 구분 (#5)
