+++ src/main/java/com/logilink/eureka/client/delivery/domain/model/dto/requestDto/UpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.logilink.eureka.client.delivery.domain.model.dto;
+package com.logilink.eureka.client.delivery.domain.model.dto.requestDto;
 
 import com.logilink.eureka.client.delivery.common.constants.DeliveryStatus;
 import lombok.Getter;
