+++ src/main/java/com/logilink/eureka/client/delivery/domain/repository/DeliveryRepository.java
@@ -1,6 +1,6 @@
 package com.logilink.eureka.client.delivery.domain.repository;
 
-import com.logilink.eureka.client.delivery.domain.model.Delivery;
+import com.logilink.eureka.client.delivery.domain.model.entity.Delivery;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
